Revert "Removed int from update_b0 condition because division leads to large number."
This reverts commit 740104ddf07c3faa4eb3ea16a67f80c995ffe83f [formerly fdfe5a3a342f312367cb3cb87dc4f846dc08b8f2] [formerly 64ef20702a873176c59951b2381ea763f642f42b [formerly 7323163d5c2370e7faf3bb004981f3e8ab40862a]].


Former-commit-id: aefb077b3d4173a8b373b5b17e01f89ccfa5f4c0 [formerly bbfcd5bd787ab3f0bf2adfd895156fe9c8f82488]
Former-commit-id: 121be533d4db8136d3f2a08dee09637f9ed1b521

+++ src/ModBatsrusMethods.f90
@@ -583,8 +583,8 @@ contains
     if(DoUpdateB0)then
        ! dB0/dt term is added at the DtUpdateB0 frequency
 
-       if((Time_Simulation/DtUpdateB0) >  &
-          ((Time_Simulation - Dt*No2Si_V(UnitT_))/DtUpdateB0)) &
+       if(int(Time_Simulation/DtUpdateB0) >  &
+          int((Time_Simulation - Dt*No2Si_V(UnitT_))/DtUpdateB0)) &
           call update_b0
     end if
